Add callback parameter to all transition method
All method that may perform an async view transition like presenting a
modal receive a completion callback parameter to be called after the
transition was executed.

This allows Portal application runner to suspend / resume message
processing to avoid updating the view in the middle of the transition.

It also adds a `isPopingTopController` property to
`PortalNavigationController` which lets the Portal application renderer
decide if view updates should be skipped in the middle of a pop
transition.