Add mechanisim to dispose resource in custom controller container
Custom component controller container now expose a method to register a
callback per custom component identifier that will be called when the
container controller will be deinitialized. This callback can be used
to free any resource that was being retained in order to render the
custom component.

For example if a custom component renderer maintains copies of custom
controller in order to maintain certain view state, those controller
should be freed when the container controller is no longer part of the
view hierarchy.